\name{twl06}
\alias{twl06}
\docType{data}
\title{
TWL06 word list
}
\description{
The Official Tournament and Club Word List
}
\usage{data(twl06)}
\format{
  A list of official Scrabble Words 
}
\source{
http://norvig.com/ngrams/TWL06.txt
}
\references{
http://en.wikipedia.org/wiki/Official_Tournament_and_Club_Word_List

}
\note{
The name "Scrabble" is a trademark of Hasbro, Inc. in the United States and Canada; elsewhere, it is a trademark of Mattel.
}
\examples{
data(twl06)
sws(head(twl06))
}
\keyword{datasets}
